[LLAMA] Switch to a fair mutex for compactation
The thread performing the compactation did suffer of starvation as it
was not able to acquire the read-write latch to carry the necessary
task. I am replacing the previously used std::shared_mutex with a
custom fair shared mutex, hoping to resolve the abnormal results
seen in the aging experiments, i.e. those with updates.