[Aging2] Fix memory footprint
In the aging2 experiment, properly compute the overhead of the internal
vectors. Before the overhead was computed according to the current capacity
of the vectors, allocated in virtual memory. Whereas, the memory footprint
of the process is based on the space used in physical memory. This
commit estimates the space usage of the internal vectors according to
their size, instead of capacity, as it assumes that all of their elts
are properly stored in physical memory.