[Teseo] Explicitly set the thread affinity
It turns out that the env. vars for the OpenMP thread affinity are
completely ignored by Clang and have a global effect, that is, they
affect all threads, in GCC. With this patch, avoid using the env.
vars for OpenMP to set the thread affinity, but set it manually
through proper pthread function calls upon necessity.